move cmd to nnictl
