+++ nni/tools/nnictl/command_utils.py
