+++ nni/tools/nnictl/common_utils.py
