+++ nni/tools/nnictl/nnictl.py
