+++ nni/tools/nnictl/nnictl_utils.py
